pyproject.toml
[project]
name = "volvo-vaen"
version = "0.1.0"
description = "Volvo Vän"
authors = [
    {name = "Federico Patota", "email" = "fpatota@google.com"},
    {name = "Anastasia Manokhina", "email" = "amanokhina@google.com"}
]
dependencies = [
    "openai>=1.12.0",
    "websockets>=12.0",
    "aiosqlite>=0.19.0",
    "fastapi>=0.115.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=21.2.0",
    "thefuzz>=0.22.1",
    "python-levenshtein>=0.27.3",
    "googlemaps>=4.10.0",
]

requires-python = ">=3.12"

[dependency-groups]
dev = [
    "pytest~=8.4.2",
    "pytest-asyncio~=1.2.0",
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]

lint = [
    "ruff>=0.4.6",
    "mypy~=1.18.2",
    "codespell~=2.4.1",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.4.20250913",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "UP045"] # ignore line too long, too complex, optional types
exclude = [".venv", "references", "ui"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv", "references", "ui"]

[tool.codespell]
ignore-words-list = "wit"
skip = "uv.lock,.venv,**/*.ipynb,*.svg,*.excalidraw,references,ui"
